+++ lifelines/utils/__init__.py
@@ -1009,6 +1009,31 @@ def pass_for_numeric_dtypes_or_raise(df):
         raise TypeError("DataFrame contains nonnumeric columns: %s. Try using pandas.get_dummies to convert the non-numeric column(s) to numerical data, or dropping the column(s)." % nonnumeric_cols)
 
 
+def check_for_immediate_deaths(stop_times_events):
+    # Only used in CTV. This checks for deaths immediately, that is (0,0) lives.
+    if ((stop_times_events['start'] == stop_times_events['stop']) & (stop_times_events['stop'] == 0) & stop_times_events['event']).any():
+        raise ValueError("""The dataset provided has subjects that die on the day of entry. (0, 0)
+is not allowed in CoxTimeVaryingFitter. If suffices to add a small non-zero value to their end - example Pandas code:
+
+> df.loc[ (df[start_col] == df[stop_col]) & (df[start_col] == 0) & df[event_col], stop_col] = 0.5
+
+Alternatively, add 1 to every subjects' final end period.
+""")
+
+
+def check_for_instantaneous_events(stop_times_events):
+    if ((stop_times_events['start'] == stop_times_events['stop']) & (stop_times_events['stop'] == 0)).any():
+        warning_text = """There exist rows in your dataframe with start and stop both at time 0:
+
+        > df.loc[(df[start_col] == df[stop_col]) & (df[start_col] == 0)]
+
+        These can be safely dropped, which will improve performance.
+
+        > df = df.loc[~((df[start_col] == df[stop_col]) & (df[start_col] == 0))]
+"""
+        warnings.warn(warning_text, RuntimeWarning)
+
+
 def check_for_overlapping_intervals(df):
     # only useful for time varying coefs, after we've done
     # some index creation
@@ -1077,7 +1102,9 @@ def to_long_format(df, duration_col):
              .drop(duration_col, axis=1)
 
 
-def add_covariate_to_timeline(long_form_df, cv, id_col, duration_col, event_col, add_enum=False, overwrite=True, cumulative_sum=False, cumulative_sum_prefix="cumsum_"):
+def add_covariate_to_timeline(long_form_df, cv, id_col, duration_col, event_col,
+                              add_enum=False, overwrite=True, cumulative_sum=False,
+                              cumulative_sum_prefix="cumsum_", delay=0):
     """
     This is a util function to help create a long form table tracking subjects' covariate changes over time. It is meant
     to be used iteratively as one adds more and more covariates to track over time. If beginning to use this function, it is recommend
@@ -1172,13 +1199,17 @@ known observation. This could case null values in the resulting dataframe."
             expanded_df['enum'] = np.arange(1, n + 1)
 
         if cumulative_sum:
-            expanded_df[cv.columns] = expanded_df[cv.columns].fillna(0)
+            expanded_df[cv.columns] = expanded_df[cv.columns].ffill().fillna(0)
 
         return expanded_df.ffill()
 
     if 'stop' not in long_form_df.columns or 'start' not in long_form_df.columns:
         raise IndexError("The columns `stop` and `start` must be in long_form_df - perhaps you need to use `lifelines.utils.to_long_format` first?")
 
+    if delay < 0:
+        raise ValueError("delay parameter must be equal to or greater than 0")
+
+    cv[duration_col] += delay
     cv = cv.dropna()
     cv = cv.sort_values([id_col, duration_col])
     cvs = cv.pipe(remove_redundant_rows)\
@@ -1281,3 +1312,4 @@ class StepSizer():
 
     def next(self):
         return self.step_size
+