+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 ### Changelogs
 
+#### 0.14.4
+ - Adds `baseline_cumulative_hazard_` and `baseline_survival_` to `CoxTimeVaryingFitter`. Because of this, new prediction methods are available. 
+ - fixed a bug in `add_covariate_to_timeline` when using `cumulative_sum` with multiple columns.
+ - Added `Likelihood ratio test` to `CoxPHFitter.print_summary` and `CoxTimeVaryingFitter.print_summary`
+ - New checks in `CoxTimeVaryingFitter` that check for immediate deaths and redundant rows. 
+ - New `delay` parameter in `add_covariate_to_timeline`
+ - removed `two_sided_z_test` from `statistics`
+
 #### 0.14.3 
  - fixes a bug when subtracting or dividing two `UnivariateFitters` with labels. 
  - fixes an import error with using `CoxTimeVaryingFitter` predict methods.